feat(license): use agora order for  vps
ref: MANAGER-5669
