+++ packages/manager/apps/dedicated/client/app/config/constants.config.js
@@ -9,7 +9,9 @@ const constants = {
     UNIVERS: 'dedicated',
     URLS: {
       CZ: {
-        express_order: 'https://www.ovh.cz/order/express/#/express/',
+        express_order: 'https://www.ovh.cz/order/express/#/new/express',
+        express_order_resume:
+          'https://www.ovh.cz/order/express/#/new/express/resume',
         support: 'http://www.ovh.cz/podpora/',
         support_contact: 'http://www.ovh.cz/podpora/',
         guides: {
@@ -40,7 +42,9 @@ const constants = {
         iplbOrder: 'https://www.ovh.cz/reseni/load-balancer/',
       },
       DE: {
-        express_order: 'https://www.ovh.de/order/express/#/express/',
+        express_order: 'https://www.ovh.de/order/express/#/new/express',
+        express_order_resume:
+          'https://www.ovh.de/order/express/#/new/express/resume',
         support: 'http://www.ovh.de/support/',
         support_contact: 'http://www.ovh.de/support/',
         guides: {
@@ -78,7 +82,9 @@ const constants = {
         iplbOrder: 'https://www.ovh.de/loesungen/load-balancer/',
       },
       ES: {
-        express_order: 'https://www.ovh.es/order/express/#/express/',
+        express_order: 'https://www.ovh.es/order/express/#/new/express',
+        express_order_resume:
+          'https://www.ovh.es/order/express/#/new/express/resume',
         support: 'http://www.ovh.es/soporte/',
         support_contact: 'http://www.ovh.es/soporte/',
         guides: {
@@ -114,7 +120,9 @@ const constants = {
         iplbOrder: 'https://www.ovh.es/soluciones/load-balancer/',
       },
       FI: {
-        express_order: 'https://www.ovh-hosting.fi/order/express/#/express/',
+        express_order: 'https://www.ovh-hosting.fi/order/express/#/new/express',
+        express_order_resume:
+          'https://www.ovh-hosting.fi/order/express/#/new/express/resume',
         support: 'http://www.ovh-hosting.fi/tuki/',
         support_contact: 'http://www.ovh-hosting.fi/tuki/',
         guides: {
@@ -152,7 +160,9 @@ const constants = {
         iplbOrder: 'https://www.ovh-hosting.fi/ratkaisut/load-balancer/',
       },
       FR: {
-        express_order: 'https://www.ovh.com/fr/order/express/#/express/',
+        express_order: 'https://www.ovh.com/fr/order/express/#/new/express',
+        express_order_resume:
+          'https://www.ovh.com/fr/order/express/#/new/express/resume',
         support: 'https://www.ovh.com/fr/support/',
         support_contact: 'https://www.ovh.com/fr/support/nous-contacter/',
         guides: {
@@ -211,7 +221,9 @@ const constants = {
         iplbOrder: 'https://www.ovh.com/fr/solutions/load-balancer/',
       },
       GB: {
-        express_order: 'https://www.ovh.co.uk/order/express/#/express/',
+        express_order: 'https://www.ovh.co.uk/order/express/#/new/express',
+        express_order_resume:
+          'https://www.ovh.co.uk/order/express/#/new/express/resume',
         support: 'http://www.ovh.co.uk/support/',
         support_contact: 'http://www.ovh.co.uk/support/',
         guides: {
@@ -260,7 +272,9 @@ const constants = {
         iplbOrder: 'https://www.ovh.co.uk/solutions/load-balancer/',
       },
       IT: {
-        express_order: 'https://www.ovh.it/order/express/#/express/',
+        express_order: 'https://www.ovh.it/order/express/#/new/express',
+        express_order_resume:
+          'https://www.ovh.it/order/express/#/new/express/resume',
         support: 'http://www.ovh.it/supporto/',
         support_contact: 'http://www.ovh.it/supporto/',
         guides: {
@@ -301,7 +315,9 @@ const constants = {
         iplbOrder: 'https://www.ovh.it/soluzioni/load-balancer/',
       },
       LT: {
-        express_order: 'https://www.ovh.lt/order/express/#/express/',
+        express_order: 'https://www.ovh.lt/order/express/#/new/express',
+        express_order_resume:
+          'https://www.ovh.lt/order/express/#/new/express/resume',
         support: 'http://www.ovh.lt/pagalba/',
         support_contact: 'http://www.ovh.lt/pagalba/',
         guides: {
@@ -341,7 +357,9 @@ const constants = {
         iplbOrder: 'https://www.ovh.lt/sprendimai/load-balancer/',
       },
       NL: {
-        express_order: 'https://www.ovh.nl/order/express/#/express/',
+        express_order: 'https://www.ovh.nl/order/express/#/new/express',
+        express_order_resume:
+          'https://www.ovh.nl/order/express/#/new/express/resume',
         support: 'http://www.ovh.nl/support/',
         support_contact: 'http://www.ovh.nl/support/',
         guides: {
@@ -378,7 +396,9 @@ const constants = {
         iplbOrder: 'https://www.ovh.nl/oplossing/load-balancer/',
       },
       PL: {
-        express_order: 'https://www.ovh.pl/order/express/#/express/',
+        express_order: 'https://www.ovh.pl/order/express/#/new/express',
+        express_order_resume:
+          'https://www.ovh.pl/order/express/#/new/express/resume',
         support: 'https://www.ovh.pl/pomoc/',
         support_contact: 'https://www.ovh.pl/pomoc/',
         guides: {
@@ -419,7 +439,9 @@ const constants = {
         iplbOrder: 'https://www.ovh.pl/rozwiazania/ip-load-balancing/',
       },
       PT: {
-        express_order: 'https://www.ovh.pt/order/express/#/express/',
+        express_order: 'https://www.ovh.pt/order/express/#/new/express',
+        express_order_resume:
+          'https://www.ovh.pt/order/express/#/new/express/resume',
         support: 'https://www.ovh.pt/suporte/',
         support_contact: 'https://www.ovh.pt/suporte/',
         guides: {
@@ -460,7 +482,9 @@ const constants = {
         iplbOrder: 'https://www.ovh.pt/solucoes/load-balancer/',
       },
       IE: {
-        express_order: 'https://www.ovh.ie/order/express/#/express/',
+        express_order: 'https://www.ovh.ie/order/express/#/new/express',
+        express_order_resume:
+          'https://www.ovh.ie/order/express/#/new/express/resume',
         support: 'https://www.ovh.ie/suport/',
         support_contact: 'https://www.ovh.ie/suport/',
         guides: {
@@ -766,7 +790,9 @@ const constants = {
     UNIVERS: 'dedicated',
     URLS: {
       ASIA: {
-        express_order: 'https://ca.ovh.com/en/order/express/#/express/',
+        express_order: 'https://ca.ovh.com/asia/order/express/#/new/express',
+        express_order_resume:
+          'https://ca.ovh.com/asia/order/express/#/new/express/resume',
         support: 'http://www.ovh.co.uk/support/',
         support_contact: 'https://www.ovh.com/ca/en/support/',
         guides: {
@@ -804,7 +830,9 @@ const constants = {
         iplbOrder: 'https://www.ovh.com/world/solutions/load-balancer/',
       },
       AU: {
-        express_order: 'https://ca.ovh.com/en/order/express/#/express/',
+        express_order: 'https://ca.ovh.com/au/order/express/#/new/express',
+        express_order_resume:
+          'https://ca.ovh.com/au/order/express/#/new/express/resume',
         support: 'http://www.ovh.co.uk/support/',
         support_contact: 'https://www.ovh.com/ca/en/support/',
         guides: {
@@ -844,7 +872,9 @@ const constants = {
       },
       CA: {
         // eq to en_CA
-        express_order: 'https://ca.ovh.com/en/order/express/#/express/',
+        express_order: 'https://ca.ovh.com/en/order/express/#/new/express',
+        express_order_resume:
+          'https://ca.ovh.com/en/order/express/#/new/express/resume',
         support: 'http://www.ovh.co.uk/support/',
         support_contact: 'https://www.ovh.com/ca/en/support/',
         guides: {
@@ -884,7 +914,9 @@ const constants = {
       },
       QC: {
         // eq to fr_CA
-        express_order: 'https://ca.ovh.com/fr/order/express/#/express/',
+        express_order: 'https://ca.ovh.com/fr/order/express/#/new/express',
+        express_order_resume:
+          'https://ca.ovh.com/fr/order/express/#/new/express/resume',
         support: 'https://www.ovh.com/fr/support/',
         support_contact: 'https://www.ovh.com/ca/fr/support/',
         guides: {
@@ -923,7 +955,9 @@ const constants = {
         iplbOrder: 'https://www.ovh.com/ca/fr/solutions/load-balancer/',
       },
       SG: {
-        express_order: 'https://ca.ovh.com/en/order/express/#/express/',
+        express_order: 'https://ca.ovh.com/sg/order/express/#/new/express',
+        express_order_resume:
+          'https://ca.ovh.com/sg/order/express/#/new/express/resume',
         support: 'http://www.ovh.co.uk/support/',
         support_contact: 'https://www.ovh.com/ca/en/support/',
         guides: {
@@ -961,7 +995,9 @@ const constants = {
         iplbOrder: 'https://www.ovh.com/world/solutions/load-balancer/',
       },
       WE: {
-        express_order: 'https://ca.ovh.com/en/order/express/#/express/',
+        express_order: 'https://us.ovh.com/us/order/express/#/new/express',
+        express_order_resume:
+          'https://us.ovh.com/us/order/express/#/new/express/resume',
         support: 'http://www.ovh.co.uk/support/',
         support_contact: 'https://www.ovh.com/ca/en/support/',
         guides: {
@@ -1000,7 +1036,9 @@ const constants = {
       },
       WS: {
         // eq to es_US
-        express_order: 'https://ca.ovh.com/en/order/express/#/express/',
+        express_order: 'https://us.ovh.com/es/order/express/#/new/express',
+        express_order_resume:
+          'https://us.ovh.com/es/order/express/#/new/express/resume',
         support: 'https://www.ovh.com/fr/support/',
         support_contact: 'https://www.ovh.com/ca/en/support/',
         guides: {
@@ -1147,7 +1185,9 @@ const constants = {
     UNIVERS: 'dedicated',
     URLS: {
       US: {
-        express_order: 'https://us.ovhcloud.com/order/express/#/express/',
+        express_order: 'https://us.ovhcloud.com/order/express/#/express',
+        express_order_resume:
+          'https://us.ovhcloud.com/order/express/#/express/review',
         support: 'http://www.ovh.co.uk/support/',
         support_contact: 'https://www.ovh.com/ca/en/support/',
         guides: {
