+++ packages/manager/apps/dedicated/client/app/license/license.module.js
@@ -1,1 +1,5 @@
-angular.module('Module.license', ['ngOvhUtils', 'ui.bootstrap']);
+angular.module('Module.license', [
+  'ngOvhUtils',
+  'ui.bootstrap',
+  'ovhManagerFilters',
+]);
