+++ packages/manager/modules/filters/src/duration-iso8601/translations/Messages_fr_FR.json
@@ -1,0 +1,5 @@
+{
+  "duration_day": "{{duration}} jours",
+  "duration_month": "{{duration}} mois",
+  "duration_year": "{{duration}} ans"
+}
