+++ packages/manager/modules/filters/src/index.js
@@ -1,9 +1,10 @@
 import angular from 'angular';
 
 import bandwidth from './bandwidth';
+import durationIso8601 from './duration-iso8601';
 
 const moduleName = 'ovhManagerFilters';
 
-angular.module(moduleName, [bandwidth]);
+angular.module(moduleName, [bandwidth, durationIso8601]);
 
 export default moduleName;
